tennistat/requirements.txt
django
django-extensions
#pygal
#django-registration-redux
#django-crispy-forms
#celery
django-filter
djangorestframework
#django-celery-results
#numpy
#matplotlib
Pillow
pandas
#git+https://github.com/Zulko/moviepy.git
# moviepy
#django-paypal

